+++ tests/testthat/test-errors.R
@@ -158,7 +158,7 @@ test_that("categorical with random effects fails", {
   expect_error(
     dynamite(obs(y ~ x + random(~1), family = "categorical") + random_spec(),
       data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-      "id", "time",
+      "time", "id",
       debug = list(no_compile = TRUE)
     ),
     "Random effects are not \\(yet\\) supported for categorical responses."
@@ -200,7 +200,7 @@ test_that("noncentered definition throws error if not of correct length", {
     NA
   )
   expect_error(
-    dynamite(obs_all_alpha, test_data, "group", "time"),
+    dynamite(obs_all_alpha, test_data, "time", "group"),
     paste(
       "Length of the `noncentered` argument of `splines\\(\\)` function",
       "is not equal to 1 or 6, the number of the channels\\."
@@ -220,7 +220,7 @@ test_that("lb_tau definition throws error if not of correct length", {
     NA
   )
   expect_error(
-    dynamite(obs_all_alpha, test_data, "group", "time"),
+    dynamite(obs_all_alpha, test_data, "time", "group"),
     paste(
       "Length of the `lb_tau` argument of `splines\\(\\)` function is not",
       "equal to 1 or 6, the number of the channels\\."
@@ -242,10 +242,12 @@ test_that("pure deterministic formula to dynamite fails", {
 
 test_that("categorical latent factor fails", {
   expect_error(
-    dynamite(obs(y ~ x, family = "categorical") + lfactor(),
-             data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "categorical") + lfactor(),
+      data = data.frame(y = factor(1:4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "No valid responses for latent factor component:\n",
@@ -256,18 +258,22 @@ test_that("categorical latent factor fails", {
 
 test_that("Latent factor errors with invalid responses", {
   expect_error(
-    dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = 1),
-             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "gaussian") + lfactor(responses = 1),
+      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     "Argument `responses` must be a <character> vector\\."
   )
   expect_error(
-    dynamite(obs(y ~ x, family = "gaussian") + lfactor(responses = "x"),
-             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "gaussian") + lfactor(responses = "x"),
+      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Argument `responses` of `lfactor\\(\\)` contains variable `x`:\n",
@@ -278,10 +284,12 @@ test_that("Latent factor errors with invalid responses", {
 test_that("Latent factor errors with nonlogical value for argument
   noncentered", {
     expect_error(
-      dynamite(obs(y ~ x, family = "gaussian") + lfactor(noncentered_lambda = 1),
-               data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-               "id", "time",
-               debug = list(no_compile = TRUE)
+      dynamite(
+        obs(y ~ x, family = "gaussian") + lfactor(noncentered_lambda = 1),
+        data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+        time = "time",
+        group = "id",
+        debug = list(no_compile = TRUE)
       ),
       "Argument `noncentered_lambda` must be a <logical> vector\\."
     )
@@ -289,10 +297,12 @@ test_that("Latent factor errors with nonlogical value for argument
 test_that("Latent factor errors with nonlogical value for argument
   nonzero_lambda", {
     expect_error(
-      dynamite(obs(y ~ x, family = "gaussian") + lfactor(nonzero_lambda = 1),
-               data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-               "id", "time",
-               debug = list(no_compile = TRUE)
+      dynamite(
+        obs(y ~ x, family = "gaussian") + lfactor(nonzero_lambda = 1),
+        data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+        time = "time",
+        group = "id",
+        debug = list(no_compile = TRUE)
       ),
       "Argument `nonzero_lambda` must be a <logical> vector\\."
     )
@@ -300,20 +310,24 @@ test_that("Latent factor errors with nonlogical value for argument
 test_that("Latent factor errors with nonlogical value for argument
   noncentered_psi", {
     expect_error(
-      dynamite(obs(y ~ x, family = "gaussian") + lfactor(noncentered_psi = 1),
-               data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-               "id", "time",
-               debug = list(no_compile = TRUE)
+      dynamite(
+        obs(y ~ x, family = "gaussian") + lfactor(noncentered_psi = 1),
+        data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+        time = "time",
+        group = "id",
+        debug = list(no_compile = TRUE)
       ),
       "Argument `noncentered_psi` must be a single <logical> value\\."
     )
   })
 test_that("Latent factor errors with nonlogical value for argument correlated", {
   expect_error(
-    dynamite(obs(y ~ x, family = "gaussian") + lfactor(correlated = 1),
-             data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
-             "id", "time",
-             debug = list(no_compile = TRUE)
+    dynamite(
+      obs(y ~ x, family = "gaussian") + lfactor(correlated = 1),
+      data = data.frame(y = rnorm(4), x = runif(4), id = 1, time = 1:4),
+      time = "time",
+      group = "id",
+      debug = list(no_compile = TRUE)
     ),
     "Argument `correlated` must be a single <logical> value\\."
   )
@@ -388,7 +402,7 @@ test_that("group variable not in data fails", {
   expect_error(
     dynamite(
       dformula = obs_test,
-      data = data.frame(y = 1, x = 1), group = "z", time = "x"
+      data = data.frame(y = 1, x = 1), time = "x", group = "z"
     ),
     "Can't find grouping variable `z` in `data`\\."
   )
@@ -398,7 +412,8 @@ test_that("time variable not in data fails", {
   expect_error(
     dynamite(
       dformula = obs_test,
-      data = data.frame(y = 1, x = 1), time = "z"
+      data = data.frame(y = 1, x = 1),
+      time = "z"
     ),
     "Can't find time index variable `z` in `data`\\."
   )
@@ -407,8 +422,10 @@ test_that("time variable not in data fails", {
 test_that("single time point fails", {
   expect_error(
     dynamite(
-      dformula = obs_test, data = data.frame(y = 1, x = 1, z = 1),
-      group = "x", time = "z"
+      dformula = obs_test,
+      data = data.frame(y = 1, x = 1, z = 1),
+      time = "z",
+      group = "x"
     ),
     "There must be at least two time points in the data."
   )
@@ -424,8 +441,8 @@ test_that("duplicated time points fail", {
         x = gl(3, 3),
         z = c(1, 2, 2, 1, 2, 3, 1, 3, 3)
       ),
-      group = "x",
       time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -456,7 +473,8 @@ test_that("missing lag variable fails", {
     dynamite(
       dformula = obs(y ~ lag(d, 1), family = "gaussian"),
       data = data.frame(y = c(1, 1), x = c(1, 1), z = c(1, 2)),
-      group = "x", time = "z",
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -471,7 +489,8 @@ test_that("missing predictor fails", {
     dynamite(
       dformula = obs(y ~ w, family = "gaussian"),
       data = data.frame(y = c(1, 1), x = c(1, 1), z = c(1, 2)),
-      group = "x", time = "z",
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     "Can't find variable `w` in `data`\\."
@@ -484,7 +503,8 @@ test_that("invalid deterministic channel definition fails", {
       dformula = obs(y ~ x, family = "gaussian") +
         aux(integer(d) ~ 1 + w),
       data = data.frame(y = c(1, 1), x = c(1, 1), z = c(1, 2)),
-      group = "x", time = "z",
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -516,7 +536,9 @@ test_that("invalid column types fail", {
   expect_error(
     dynamite(
       dformula = obs(y ~ x, family = "gaussian"),
-      data = test_data, group = "x", time = "z",
+      data = test_data,
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -534,7 +556,9 @@ test_that("non-finite values in data fail", {
   expect_error(
     dynamite(
       dformula = obs(y ~ x, family = "gaussian"),
-      data = test_data, group = "x", time = "z",
+      data = test_data,
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     "Non-finite values were found in variables `y`, `w`, and `u` of `data`\\."
@@ -546,7 +570,9 @@ test_that("non-factor categorical response fails", {
   expect_error(
     dynamite(
       dformula = obs(y ~ 1, family = "categorical"),
-      data = test_data, group = "x", time = "z",
+      data = test_data,
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -567,7 +593,9 @@ test_that("factor types for non-categorical families fails", {
     expect_error(
       dynamite(
         dformula = obs(form, family = f),
-        data = test_data, group = "x", time = "z",
+        data = test_data,
+        time = "z",
+        group = "x",
         debug = list(no_compile = TRUE)
       ),
       paste0(
@@ -586,7 +614,9 @@ test_that("negative values for distributions with positive support fails", {
     expect_error(
       dynamite(
         dformula = obs(form, family = f),
-        data = test_data, group = "x", time = "z",
+        data = test_data,
+        time = "z",
+        group = "x",
         debug = list(no_compile = TRUE)
       ),
       paste0(
@@ -602,7 +632,9 @@ test_that("bernoulli without 0/1 values fails", {
   expect_error(
     dynamite(
       dformula = obs(y ~ 1, family = "bernoulli"),
-      data = test_data, group = "x", time = "z",
+      data = test_data,
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -617,7 +649,9 @@ test_that("beta without (0, 1) values fails", {
   expect_error(
     dynamite(
       dformula = obs(y ~ 1, family = "beta"),
-      data = test_data, group = "x", time = "z",
+      data = test_data,
+      time = "z",
+      group = "x",
       debug = list(no_compile = TRUE)
     ),
     paste0(
@@ -942,9 +976,13 @@ f <- obs(y ~ -1 + random(~1) + z + varying(~ x + lag(y)), family = "gaussian") +
 test_that("incomplete priors fails", {
   p2 <- p[-1, ]
   expect_error(
-    dynamite(f,
-             data = gaussian_example, time = "time", group = "id",
-             priors = p2, debug = list(no_compile = TRUE)
+    dynamite(
+      f,
+      data = gaussian_example,
+      time = "time",
+      group = "id",
+      priors = p2,
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Argument `priors` must contain all relevant parameters:\n",
@@ -962,9 +1000,13 @@ test_that("irrevelant parameters fails", {
     category = ""
   ))
   expect_error(
-    dynamite(f,
-             data = gaussian_example, time = "time", group = "id",
-             priors = p2, debug = list(no_compile = TRUE)
+    dynamite(
+      f,
+      data = gaussian_example,
+      time = "time",
+      group = "id",
+      priors = p2,
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Argument `priors` must contain only relevant parameters:\n",
@@ -977,9 +1019,13 @@ test_that("irrevelant parameters fails", {
 test_that("unsupported prior distribution fails", {
   p$prior[5] <- "aaa"
   expect_error(
-    dynamite(f,
-             data = gaussian_example, time = "time", group = "id",
-             priors = p, debug = list(no_compile = TRUE)
+    dynamite(
+      f,
+      data = gaussian_example,
+      time = "time",
+      group = "id",
+      priors = p,
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Found an unsupported prior distribution in `priors`:\n",
@@ -991,9 +1037,13 @@ test_that("unsupported prior distribution fails", {
 test_that("constrained prior for unconstrained parameter fails", {
   p$prior[5] <- "gamma(2, 1)"
   expect_error(
-    dynamite(f,
-             data = gaussian_example, time = "time", group = "id",
-             priors = p, debug = list(no_compile = TRUE)
+    dynamite(
+      f,
+      data = gaussian_example,
+      time = "time",
+      group = "id",
+      priors = p,
+      debug = list(no_compile = TRUE)
     ),
     paste0(
       "Priors for parameters alpha, beta, and delta ",